Fail when common ancestor container is root
Based on the following document:
"If no element can be found in the ancestry of the Range that has a non-empty id attribute and is contained by the root then this function will raise an exception."
The root node should not be considered as a valid common ancestor container. But no error is thrown in this scenario.